Expand tournament structure
Introduce enum-backed lookup types and a JSON-based stage configuration, plus various API and model cleanups.

Details:
- pom: add Lombok to annotationProcessorPaths for compilation.
- New domain types: RankingMethod, StageMode, FinalRankingStrategy and their Id enums; DocumentTypeId and StatusId enums.
- Replace several previous numeric/PK lookup tables with enum IDs (DocumentType, Status) and remove their repositories/resources.
- Add StageConfig class and map stage/ stageTemplate config to JSONB (Stage.stageConfig / StageTemplate.stageConfig). Remove StageOption entity and related repo/resource.
- Add StageMode and RankingMethod relations to Stage and StageTemplate.
- Add OpenAPI @Tag annotations to many REST resources for clearer API docs.
- Add timestamp management (@CreationTimestamp / @UpdateTimestamp) to entities (Match, Round, Set, TournamentRegistration) and default values for boolean fields where appropriate.
- Change several numeric types to smaller types (Long -> Short) for counts/totals and adjust some column definitions to use TEXT for large descriptions/notes.
- Add setup-defaults.sql and minor repository/resource formatting fixes.

These changes simplify lookup data handling (using enums), introduce structured JSON stage settings, and improve auditing and API documentation.

+++ src/main/java/dev/laubfrosch/archery/backend/participant/Team.java
@@ -35,5 +35,5 @@ public class Team extends PanacheEntityBase {
     private String contactEmail;
 
     @Column(name = "team_member_count")
-    private Long teamMemberCount;
+    private Short teamMemberCount;
 }
